moved ConnectDialog into separate package and decompose nested classes
